// default-frontend/app/types/openapi.d.ts
declare namespace Components {
    namespace Schemas {
        export interface BestHotelOffer {
            hotel: Hotel;
            minPrice: number;
            departureDate: string; // date
            returnDate: string; // date
            roomType?: string;
            mealType?: string;
            countAdults: number;
            countChildren: number;
            duration: number;
            countAvailableOffers: number;
        }
        export type GetBestHotelOffersResponse = BestHotelOffer[];
        export interface GetHotelOffersResponse {
            hotel: Hotel;
            items: Offer[];
        }
        export interface Hotel {
            id: number;
            name: string;
            stars: number;
        }
        export interface Offer {
            price: number;
            countAdults: number;
            countChildren: number;
            inboundDepartureAirport: string;
            inboundDepartureDatetime: string; // date
            inboundArrivalAirport: string;
            inboundArrivalDatetime: string; // datetime
            outboundDepartureAirport: string;
            outbundDepartureDatetime?: string; // date
            outboundArrivalAirport: string;
            outboundArrivalDatetime: string; // datetime
            mealType: string;
            oceanView: boolean;
            roomType: string;
        }
    }
}